各セクションをコンポーネント化
