+++ src/components/Footer.jsx
@@ -1,0 +1,13 @@
+export const Footer = () => {
+    return (
+        <>
+            <footer className="footer sm:footer-horizontal footer-center bg-base-300 text-base-content p-4">
+                <aside>
+                    <p>
+                        Copyright © {new Date().getFullYear()} - All right reserved by Kyo
+                    </p>
+                </aside>
+            </footer>
+        </>
+    );
+};
